+++ Sources/KeyboardKit/_Deprecated/AutocompleteContext+Sync.swift
@@ -1,11 +1,3 @@
-//
-//  AutocompleteContext+Sync.swift
-//  KeyboardKit
-//
-//  Created by Daniel Saidi on 2024-06-04.
-//  Copyright © 2024 Daniel Saidi. All rights reserved.
-//
-
 import Foundation
 import SwiftUI
 
@@ -13,7 +5,9 @@ public extension AutocompleteContext {
 
     /// DEPRECATED!
     ///
-    /// > Warning: Settings has been moved to ``AutocompleteContext``. This function will be removed in KeyboardKit 9.0.
+    /// > Warning: Settings has been moved to this type.
+    ///
+    /// This function will be removed in KeyboardKit 9.0.
     func sync(with settings: AutocompleteSettings) {
         DispatchQueue.main.async {
             self.syncAfterAsync(with: settings)