+++ Sources/KeyboardKit/_Deprecated/Keyboard+Settings.swift
@@ -10,28 +10,10 @@ import SwiftUI
 
 public extension Keyboard {
 
-    /// This type specifies global keyboard settings.
+    /// DEPRECATED!
     ///
-    /// This lets us decouple the input view controller from
-    /// any views that require settings, services, and state
-    /// from the controller.
-    ///
-    /// Do not confuse this with ``KeyboardSettings``, which
-    /// is an observable class for general keyboard settings
-    /// that you can access via ``keyboardSettings``.
-    ///
-    /// Unlike the ``Keyboard/State`` properties, these will
-    /// persist any changes that are made to them. Make sure
-    /// to see the <doc:Essentials> & <doc:Settings-Article>
-    /// articles for more information on how to sync changes
-    /// between the main app and the keyboard extension.
-    ///
-    /// KeyboardKit will automatically sync applied settings
-    /// with ``KeyboardInputViewController/state`` when your
-    /// keyboard launches, and will add `onChange` events to
-    /// the keyboard view to apply settings changes to these
-    /// state instances as well. Context changes are however
-    /// not synced to settings.
+    /// > Warning: Settings have been moved into the various
+    /// context types. This will be removed in KeyboardKit 9.
     class Settings {
 
         /// The autocomplete settings to use.